+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add: bluetooth device-added, device-removed signal
 - add: cursor property
 - feat: window popup close on click away
+- add: config.onWindowToggled & config.onConfigParsed
 
 ## Breaking Changes